refactor: extract shared utilities to cloud_utils module
Extract common utility functions duplicated across S3, GCS, and Azure
adapters into a centralized cloud_utils module. This reduces code
duplication and improves maintainability.

Extracted utilities:
- Encoding: bytes_to_hex, base64_encode/decode, base64url_encode, url_encode
- Crypto: sha256, sha256_bytes, hmac_sha256
- Time: get_iso8601_time, get_date_stamp, get_rfc3339_time, get_rfc1123_time
- Random: generate_random_bytes, generate_random_hex
- Parsing: extract_xml_element, extract_json_value
- Content: detect_content_type
- Retry: calculate_retry_delay, is_retryable_status

Code reduction:
- Total lines reduced from 7,624 to 7,255 (369 lines, ~5%)
- Removed ~830 lines of duplicate code across adapters

Part of #155
