+++ CMakeLists.txt
@@ -290,6 +290,7 @@ endif()
 # Add cloud storage sources (requires OpenSSL for signing)
 if(FILE_TRANS_ENABLE_ENCRYPTION AND OpenSSL_FOUND)
     list(APPEND FILE_TRANS_SOURCES
+        src/cloud/cloud_utils.cpp
         src/cloud/s3_storage.cpp
         src/cloud/azure_blob_storage.cpp
         src/cloud/gcs_storage.cpp
